sp/requirements.txt
# StillPoint Desktop and Server requirements (no PySide UI)

PySide6>=6.6
fastapi>=0.110
python-multipart>=0.0.6
uvicorn>=0.29
httpx>=0.27
watchdog>=4.0
pydantic>=2.7
pytest>=8.2
python-jose[cryptography]>=3.3
argon2-cffi>=23.1
passlib>=1.7

chromadb>=0.3.26
pdfminer.six>=20221105
python-docx>=0.8.11
pytesseract>=0.3.10
Pillow>=10.0

keyboard>=0.13
markdown>=3.5
Pygments>=2.18
Flask>=3.0
Jinja2>=3.1
